spark/requirements.txt
delta-spark==3.2.1
loguru==0.7.2
pydantic==2.9.2
requests==2.32.3
